# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ArMenu.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/platform/businesses": {
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "Every business on the platform, by name.",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "Opens a business with its owner's account, who signs in with the given user name and password.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenBusinessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenedBusinessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/platform/businesses/{businessId}/enter": {
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "An access token for the business with the owner's role. Short-lived and not refreshable: enter again when it expires.",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnteredBusinessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusinessSummary": {
        "required": [
          "id",
          "name",
          "slug",
          "status",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "EnteredBusinessResponse": {
        "required": [
          "accessToken",
          "expiresAt",
          "expiresIn",
          "workspace"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          },
          "workspace": {
            "type": "string"
          },
          "tokenType": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "A bearer token for a business the administrator entered, and the business's slug. There is no refresh token:\nwhen it expires, the dashboard asks to enter again."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. `tenant.not_found`."
          },
          "traceId": {
            "type": "string",
            "description": "Correlates the response with server logs and traces."
          },
          "errorCodes": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Stable error codes per field (JSON path), for localized messages on the client."
          }
        }
      },
...
      "OpenBusinessRequest": {
        "required": [
          "businessName",
          "slug",
          "defaultCulture",
          "currency",
          "ownerFullName",
          "ownerUserName",
          "ownerPassword",
          "timeZone"
        ],
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "defaultCulture": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "ownerFullName": {
            "type": "string"
          },
          "ownerUserName": {
            "type": "string"
          },
          "ownerPassword": {
            "type": "string"
          },
          "timeZone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "OpenedBusinessResponse": {
        "required": [
          "id",
          "slug"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. `tenant.not_found`."
          },
          "traceId": {
            "type": "string",
            "description": "Correlates the response with server logs and traces."
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token from sign-in or refresh. Tokens are scoped to one tenant.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}